+++ src/modules/financials/expedition-paid-date.in.ts
@@ -1,0 +1,42 @@
+import { IsDate, IsNumber } from 'class-validator';
+import { Field, InputType, Int } from 'type-graphql';
+import { ExpeditionPaidDate } from './entities/expedition-paid-date';
+
+@InputType()
+export class ExpeditionPaidDateInput implements Partial<ExpeditionPaidDate> {
+  @Field(() => Int)
+  @IsNumber()
+  id!: number;
+
+  @Field({ nullable: true })
+  @IsDate()
+  circulationDate?: Date;
+
+  @Field({ nullable: true })
+  expeditionName?: string;
+
+  @Field({ nullable: true })
+  invoiceNo?: string;
+
+  @Field({ nullable: true })
+  poNumber?: string;
+
+  @Field({ nullable: true })
+  ajuNo?: string;
+
+  @Field({ nullable: true })
+  totalInvoice?: number;
+
+  @Field({ nullable: true })
+  @IsDate()
+  paidDate?: Date;
+
+  @Field({ defaultValue: false })
+  isPaid?: boolean;
+
+  @Field({ nullable: true })
+  notes?: string;
+
+  @Field({ nullable: true })
+  plant?: string;
+}